Обляшевский С.А. P32312

+++ solution/src/bmp_reader.c
@@ -9,7 +9,7 @@ enum read_status read_bmp_file(FILE* input, struct image* image){
     if (head_read_flag != 1) return READ_ERROR_HEADING;
     uint64_t width = header.biWidth;
     uint64_t height = header.biHeight;
-    uint64_t padding = (width*height*sizeof(struct pixel))%4;
+    uint8_t padding = (width*height*sizeof(struct pixel))%4;
     if (padding) padding = 4 - padding;
     create_image(width, height);
     for (uint64_t i = 0; i < height; i++){
